refactor json payload
